Update cds extension tests.
