+++ tests/test_cds_extension/test_archive.py
@@ -1,13 +1,15 @@
 import unittest
 
-from tests import CdsTestMixin, assert_archive_created
+from tests import CdsTestMixin
+
+from . import assert_archive_created
 
 
 class ArchiveTest(CdsTestMixin, unittest.TestCase):
     @assert_archive_created
     def test_create_archive(self):
-        self.assert_python_source_ok(f'import _cds; _cds._create_archive("{self.TEST_ARCHIVE}")')
-        self.assert_python_source_ok(f'import _cds; _cds._load_archive("{self.TEST_ARCHIVE}")')
+        self.assert_python_source_ok(f'import _cds; _cds._create_archive("{self.TEST_ARCHIVE}")', PYCDSMODE='MANUALLY')
+        self.assert_python_source_ok(f'import _cds; _cds._load_archive("{self.TEST_ARCHIVE}")', PYCDSMODE='MANUALLY')
 
     @assert_archive_created
     def test_create_multiple_archive(self):
@@ -19,7 +21,7 @@ with pytest.raises(_cds.CDSException) as e:
     _cds._create_archive("{self.TEST_ARCHIVE}")
     _cds._create_archive("{self.TEST_ARCHIVE}")
 e.match("cds already initialized.")
-""")
+""", PYCDSMODE='MANUALLY')
 
     @assert_archive_created
     def test_dump_not_supported_type(self):
@@ -41,7 +43,7 @@ import _cds
 with pytest.raises(_cds.CDSException) as e:
     _cds._load_archive("{self.TEST_ARCHIVE}")
 e.match("open mmap file failed.")
-""")
+""", PYCDSMODE='MANUALLY')
 
     @assert_archive_created
     def test_load_null(self):
@@ -71,4 +73,4 @@ t1 = _cds._get_obj()
 t2 = _cds._get_obj()
 assert t1 == t2
 assert id(t1) == id(t2)
-""")
+""", PYCDSMODE='MANUALLY')
